step by step implementation for flet todo app
