--- todo/todo.py
@@ -1,171 +1,0 @@
-import flet as ft
-
-class Task(ft.Column):
-    def __init__(self, task_name, task_status_change, task_delete):
-        super().__init__()
-        self.completed = False
-        self.task_name = task_name
-        self.task_status_change = task_status_change
-        self.task_delete = task_delete
-        self.display_task = ft.Checkbox(
-            value=False, label=self.task_name, on_change=self.status_changed
-        )
-        self.edit_name = ft.TextField(expand=1)
-
-        self.display_view = ft.Row(
-            alignment=ft.MainAxisAlignment.SPACE_BETWEEN,
-            vertical_alignment=ft.CrossAxisAlignment.CENTER,
-            controls=[
-                self.display_task,
-                ft.Row(
-                    spacing=0,
-                    controls=[
-                        ft.IconButton(
-                            icon=ft.icons.CREATE_OUTLINED,
-                            tooltip="Edit To-Do",
-                            on_click=self.edit_clicked,
-                        ),
-                        ft.IconButton(
-                            ft.icons.DELETE_OUTLINE,
-                            tooltip="Delete To-Do",
-                            on_click=self.delete_clicked,
-                        ),
-                    ],
-                ),
-            ],
-        )
-
-        self.edit_view = ft.Row(
-            visible=False,
-            alignment=ft.MainAxisAlignment.SPACE_BETWEEN,
-            vertical_alignment=ft.CrossAxisAlignment.CENTER,
-            controls=[
-                self.edit_name,
-                ft.IconButton(
-                    icon=ft.icons.DONE_OUTLINE_OUTLINED,
-                    icon_color=ft.colors.GREEN,
-                    tooltip="Update To-Do",
-                    on_click=self.save_clicked,
-                ),
-            ],
-        )
-        self.controls = [self.display_view, self.edit_view]
-
-    def edit_clicked(self, e):
-        self.edit_name.value = self.display_task.label
-        self.display_view.visible = False
-        self.edit_view.visible = True
-        self.update()
-
-    def save_clicked(self, e):
-        self.display_task.label = self.edit_name.value
-        self.display_view.visible = True
-        self.edit_view.visible = False
-        self.update()
-
-    def status_changed(self, e):
-        self.completed = self.display_task.value
-        self.task_status_change(self)
-
-    def delete_clicked(self, e):
-        self.task_delete(self)
-
-
-class TodoApp(ft.Column):
-    # application's root control is a Column containing all other controls
-    def __init__(self):
-        super().__init__()
-        self.new_task = ft.TextField(
-            hint_text="What needs to be done?", on_submit=self.add_clicked, expand=True
-        )
-        self.tasks = ft.Column()
-
-        self.filter = ft.Tabs(
-            scrollable=False,
-            selected_index=0,
-            on_change=self.tabs_changed,
-            tabs=[ft.Tab(text="all"), ft.Tab(text="active"), ft.Tab(text="completed")],
-        )
-
-        self.items_left = ft.Text("0 items left")
-
-        self.width = 600
-        self.controls = [
-            ft.Row(
-                [ft.Text(value="Todos", theme_style=ft.TextThemeStyle.HEADLINE_MEDIUM)],
-                alignment=ft.MainAxisAlignment.CENTER,
-            ),
-            ft.Row(
-                controls=[
-                    self.new_task,
-                    ft.FloatingActionButton(
-                        icon=ft.icons.ADD, on_click=self.add_clicked
-                    ),
-                ],
-            ),
-            ft.Column(
-                spacing=25,
-                controls=[
-                    self.filter,
-                    self.tasks,
-                    ft.Row(
-                        alignment=ft.MainAxisAlignment.SPACE_BETWEEN,
-                        vertical_alignment=ft.CrossAxisAlignment.CENTER,
-                        controls=[
-                            self.items_left,
-                            ft.OutlinedButton(
-                                text="Clear completed", on_click=self.clear_clicked
-                            ),
-                        ],
-                    ),
-                ],
-            ),
-        ]
-
-    def add_clicked(self, e):
-        if self.new_task.value:
-            task = Task(self.new_task.value, self.task_status_change, self.task_delete)
-            self.tasks.controls.append(task)
-            self.new_task.value = ""
-            self.new_task.focus()
-            self.update()
-
-    def task_status_change(self, task):
-        self.update()
-
-    def task_delete(self, task):
-        self.tasks.controls.remove(task)
-        self.update()
-
-    def tabs_changed(self, e):
-        self.update()
-
-    def clear_clicked(self, e):
-        for task in self.tasks.controls[:]:
-            if task.completed:
-                self.task_delete(task)
-
-    def before_update(self):
-        status = self.filter.tabs[self.filter.selected_index].text
-        count = 0
-        for task in self.tasks.controls:
-            task.visible = (
-                status == "all"
-                or (status == "active" and task.completed == False)
-                or (status == "completed" and task.completed)
-            )
-            if not task.completed:
-                count += 1
-        self.items_left.value = f"{count} active item(s) left"
-
-
-def main(page: ft.Page):
-    page.title = "ToDo App"
-    page.horizontal_alignment = ft.CrossAxisAlignment.CENTER
-    page.scroll = ft.ScrollMode.ADAPTIVE
-
-    # create app control and add it to the page
-    page.add(TodoApp())
-
-
-ft.app(main)
