+++ todo/todo_5.py
@@ -1,0 +1,61 @@
+import flet as ft
+
+class TodoApp(ft.Column):
+    # application's root control is a Column containing all other controls
+    def __init__(self):
+        super().__init__()
+        self.new_task = ft.TextField(hint_text="What needs to be done?", expand=True)
+        self.tasks_view = ft.Column()
+
+        self.filter = ft.Tabs(
+            selected_index=0,
+            on_change=self.tabs_changed,
+            tabs=[ft.Tab(text="all"), ft.Tab(text="active"), ft.Tab(text="completed")],
+        )
+
+        self.width = 600
+        self.controls = [
+            ft.Row(
+                controls=[
+                    self.new_task,
+                    ft.FloatingActionButton(
+                        icon=ft.icons.ADD, on_click=self.add_clicked
+                    ),
+                ],
+            ),
+            self.tasks_view,
+        ]
+
+    def add_clicked(self, e):
+        self.tasks_view.controls.append(ft.Checkbox(label=self.new_task.value))
+        self.new_task.value = ""
+        self.update()
+
+    def before_update(self):
+        status = self.filter.tabs[self.filter.selected_index].text
+        for task in self.tasks_view.controls:
+            task.visible = (
+                status == "all"
+                or (status == "active" and task.completed == False)
+                or (status == "completed" and task.completed)
+            )
+
+    def tabs_changed(self, e):
+        self.update()
+
+    def task_status_change(self, e):
+        self.update()
+
+
+def main(page: ft.Page):
+    page.title = "To-Do App"
+    page.horizontal_alignment = ft.CrossAxisAlignment.CENTER
+    page.update()
+
+    # create application instance
+    todo = TodoApp()
+
+    # add application's root control to the page
+    page.add(todo)
+
+ft.app(main)
